WIP - Segment refactoring - Move relative dates to separate class
