+++ app/bundles/LeadBundle/Segment/RelativeDate.php
@@ -1,0 +1,62 @@
+<?php
+
+/*
+ * @copyright   2014 Mautic Contributors. All rights reserved
+ * @author      Mautic
+ *
+ * @link        http://mautic.org
+ *
+ * @license     GNU/GPLv3 http://www.gnu.org/licenses/gpl-3.0.html
+ */
+
+namespace Mautic\LeadBundle\Segment;
+
+use Symfony\Component\Translation\TranslatorInterface;
+
+class RelativeDate
+{
+    /** @var TranslatorInterface */
+    private $translator;
+
+    public function __construct(TranslatorInterface $translator)
+    {
+        $this->translator = $translator;
+    }
+
+    /**
+     * @return array
+     */
+    public function getRelativeDateStrings()
+    {
+        $keys = $this->getRelativeDateTranslationKeys();
+
+        $strings = [];
+        foreach ($keys as $key) {
+            $strings[$key] = $this->translator->trans($key);
+        }
+
+        return $strings;
+    }
+
+    /**
+     * @return array
+     */
+    private function getRelativeDateTranslationKeys()
+    {
+        return [
+            'mautic.lead.list.month_last',
+            'mautic.lead.list.month_next',
+            'mautic.lead.list.month_this',
+            'mautic.lead.list.today',
+            'mautic.lead.list.tomorrow',
+            'mautic.lead.list.yesterday',
+            'mautic.lead.list.week_last',
+            'mautic.lead.list.week_next',
+            'mautic.lead.list.week_this',
+            'mautic.lead.list.year_last',
+            'mautic.lead.list.year_next',
+            'mautic.lead.list.year_this',
+            'mautic.lead.list.anniversary',
+        ];
+    }
+}
